Optimize shooting plant speed by reducing bulletInterval to 1.5s and adjusting animation frame rates for Peashooter, Snow Pea, and Repeater

+++ Repeater.h
@@ -4,7 +4,7 @@
 
 class Repeater : public ShootingPlant
 {
-	int bulletinterval_2;
+	float bulletinterval_2;
 	float bulletTimer_2;
 public : 
 	Repeater();
